refactor: modify folder stuct and make css to scss
