+++ package.json
@@ -50,6 +50,7 @@
     "optimize-css-assets-webpack-plugin": "^4.0.1",
     "postcss-flexbugs-fixes": "3.3.1",
     "postcss-loader": "2.1.5",
+    "postcss-svg": "^2.4.0",
     "prettier": "^1.13.5",
     "promise": "8.0.1",
     "prop-types": "^15.6.2",
