+++ src/pages/App.js
@@ -1,18 +1,20 @@
 import React, { Component } from 'react'
-import logo from './logo.svg'
-import './App.css'
+import classNames from 'classnames/bind'
+import logo from '../assets/logo.svg'
+import styles from './App.module.scss'
 
+const cx = classNames.bind(styles)
 class App extends Component {
   render() {
     return (
-      <div className="App">
-        <header className="App-header">
-          <img src={logo} className="App-logo" alt="logo" />
+      <div className={cx('app')}>
+        <header className={cx('app-header')}>
+          <img src={logo} className={cx('app-logo')} alt="logo" />
           <p>
             Edit <code>src/App.js</code> and save to reload.
           </p>
           <a
-            className="App-link"
+            className={cx('app-link')}
             href="https://reactjs.org"
             target="_blank"
             rel="noopener noreferrer">
