+++ src/pages/App.module.scss
@@ -1,28 +1,29 @@
-.App {
+.app {
   text-align: center;
+  background: #aaa;
 }
 
-.App-logo {
+.app-logo {
   height: 40vmin;
-  animation: App-logo-spin infinite 20s linear;
+  animation: app-logo-spin infinite 20s linear;
 }
 
-.App-header {
+.app-header {
   display: flex;
   flex-direction: column;
   align-items: center;
   justify-content: center;
   min-height: 100vh;
   font-size: calc(10px + 2vmin);
-  color: white;
+  color: #fff;
   background-color: #282c34;
 }
 
-.App-link {
+.app-link {
   color: #61dafb;
 }
 
-@keyframes App-logo-spin {
+@keyframes app-logo-spin {
   from {
     transform: rotate(0deg);
   }
